chore: :art: :truck: moves common exceptions to specific folder
