+++ src/main/java/com/college/hotlittlepigs/exception/common/JwtException.java
@@ -1,0 +1,10 @@
+package com.college.hotlittlepigs.exception.common;
+
+public class JwtException extends RuntimeException {
+
+  private static final long serialVersionUID = 1L;
+
+  public JwtException(String msg) {
+    super(msg);
+  }
+}
